Applied js_min fix.

+++ bin/jsminify.sh
@@ -1,6 +1,7 @@
 js_minify() {	
 	local demandwareServer=$(read_conf "deploy" "demandwareServer" $demandwareServer)
 	local minifyJS=$(prompt $demandwareServer "minifyJS" $bool false "Would you like to minify all Javascript files" true "" true)
+	local skipCartridges=false
 	echo
 
 	if [ $minifyJS == true ]; then
@@ -13,49 +14,48 @@ js_minify() {
 			if [[ "$installMinify" == "y" || "$installMinify" == "Y" ]]; then
 				gem install uglifier
 			else
-				exit 1
+				skipCartridges=true
 			fi
 		fi
 
 		echo "Minifying .js files."
 		echo
 
-		for cartridge in "${cartridges[@]}"; do
-			echo
-			local cartPath=$(scp_cartridge_path "$cartridge")
-			cd ${homedir}/${cartPath}
-			echo ${cartridge}
+		if [ $skipCartridges == false ]; then
+			for cartridge in "${cartridges[@]}"; do
+				echo
+				local cartPath=$(scp_cartridge_path "$cartridge")
+				cd ${homedir}/${cartPath}
+				echo ${cartridge}
 
-			if [ -d ${homedir}/${cartPath}/cartridge/static ]; then
+				if [ -d ${homedir}/${cartPath}/cartridge/static ]; then
 
-				file_list=()
-				while IFS= read -d $'\0' -r file; do
-					file_list=("${file_list[@]}" "$file")
-				done < <(find "${homedir}/${cartPath}/cartridge/static" -name "*.js" ! -name '*.min.js' -print0)
+					file_list=()
+					while IFS= read -d $'\0' -r file; do
+						file_list=("${file_list[@]}" "$file")
+					done < <(find "${homedir}/${cartPath}/cartridge/static" -name "*.js" ! -name '*.min.js' -print0)
 
-				cd ${deploydir}/bin
-				for file in "${file_list[@]}" ; do
-					echo "Minifying Javascript $file"
-					local jsFile=${file}
-					#if [ iscygwin ]; then
-					#	jsFile=$(cygpath -w $file)
-					#fi
+					cd ${deploydir}/bin
+					for file in "${file_list[@]}" ; do
+						echo "Minifying Javascript $file"
+						local jsFile=${file}
+						#if [ iscygwin ]; then
+						#	jsFile=$(cygpath -w $file)
+						#fi
 
-					$(which ruby) jsmin.rb $file >> ${file}.tmp
+						$(which ruby) jsmin.rb $file >> ${file}.tmp
 
-					#cp ${file} ${file}.bak
-					#java -jar yuicompressor-2.4.8.jar --type js "${jsFile}" > ${file}.tmp
-					if [[ -f ${file}.tmp && -s ${file}.tmp ]]; then
-						mv ${file}.tmp ${file}
-					else
-						rm ${file}.tmp
-					fi
-					#rm ${file}.bak
-
-				done
-
-			fi
-
-		done
+						#cp ${file} ${file}.bak
+						#java -jar yuicompressor-2.4.8.jar --type js "${jsFile}" > ${file}.tmp
+						if [[ -f ${file}.tmp && -s ${file}.tmp ]]; then
+							mv ${file}.tmp ${file}
+						else
+							rm ${file}.tmp
+						fi
+						#rm ${file}.bak
+					done
+				fi
+			done
+		fi
 	fi
 }
